backend/requirements.txt
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3
google-generativeai==0.3.0
python-dotenv==1.0.0
